+++ packages/components/organisms/f-checkout/src/components/_tests/Error.test.js
@@ -1,41 +1,33 @@
 import Vuex from 'vuex';
-import { VueI18n } from '@justeat/f-globalisation';
 import { shallowMount, createLocalVue } from '@vue/test-utils';
 import Error from '../Error.vue';
-import {
-    i18n, createStore
-} from './helpers/setup';
 
 const localVue = createLocalVue();
 
-localVue.use(VueI18n);
 localVue.use(Vuex);
 
 describe('Error', () => {
-    const propsData = {};
+    let wrapper;
+
+    beforeEach(() => {
+        // Arrange & Act
+        wrapper = shallowMount(Error, {
+            localVue,
+            propsData: {
+                header: 'Test header',
+                description: 'Test description'
+            }
+        });
+    });
+
 
     it('should be defined', () => {
-        // Arrange & Act
-        const wrapper = shallowMount(Error, {
-            store: createStore(),
-            i18n,
-            localVue,
-            propsData
-        });
-
         // Assert
         expect(wrapper.exists()).toBe(true);
     });
 
     it('should show the heading', () => {
-        // Arrange & Act
-        const wrapper = shallowMount(Error, {
-            store: createStore(),
-            i18n,
-            localVue,
-            propsData
-        });
-
+        // Act
         const heading = wrapper.find('[data-test-id="checkout-error-page-heading"]');
 
         // Assert
@@ -43,14 +35,7 @@ describe('Error', () => {
     });
 
     it('should show the description', () => {
-        // Arrange & Act
-        const wrapper = shallowMount(Error, {
-            store: createStore(),
-            i18n,
-            localVue,
-            propsData
-        });
-
+        // Act
         const description = wrapper.find('[data-test-id="checkout-error-page-description"]');
 
         // Assert